LayoutTests/platform/mac-mojave/fast/text/international/system-language/navigator-language/navigator-language-hi-expected.txt

PASS navigator.language is "hi-IN"
PASS successfullyParsed is true

TEST COMPLETE
